+++ src/index.js
@@ -1,5 +1,6 @@
-export { Provider, createCustomContext } from './Provider';
+export { createContainer } from './createContainer';
+
+export { Provider } from './Provider';
 export { useTrackedState, useTracked } from './useTrackedState';
 export { useDispatch } from './useDispatch';
 export { useSelector } from './useSelector';
-export { createContainer } from './createContainer';